Create abstract dataset for years field
Stop using Protocol for Dataset
